string_generator.py Improvements and Refactoring
Replaced random.choice() with secrets.choice().  Replaced some calls to list(), which was uneeded as .choice() can pull from a string.  Only function that needs a list of characters in shuffle(), which incidentally was renamed to _shuffle_characters as having a function paramater also named shuffle was a bad choice.