string_generator Refactor and Compatibility
string_generator.py was heavily restructured so functions and methods were better organized.  While improvements to structure were made prior to this update, some function still did too much, and getting a randomized string was still a bit of a hassle.  A new class, RandomString, handles everything related to string randomization now.  The class also defines __repr__ so instatiating an object essentially makes that object the randomized string; no outside method calls needed.

The import section was also updated to account for Python versions < 3.6 -- the use of the secrets module meant the program was limited to only those with the latest version of Python.  A try-except block now tries to import the secrets module; if this fails, random.SystemRandom() is called directly.  the secrets module already calls random.SystemRandom() to get its random data, so functionally, there shouldn't be a difference.  In either case, whatever module Python can successfully import first gets assigned to RAND_METHOD.  This allows the .choice() method to be called regardless of how SystemRandom() is called.