Better digest comparison and name changes
Hash digests are compared via hmac.compare_digest() now.  This is a more secure method of comparing two digests, and it reduces the risk of time analysis.

All references to to the word 'checksum' have been changes.  Checksum was a poor choice of words as cmpsums.py is actually comparing digests.  As such, all references to the word 'checksum' has been renamed to something more appropriate.  On that note, cmpsums.py has been renamed to hashchk.py.