pysha3 now included to support hashlib 3.6 methods in python versions < 3.6
hashlib incorporpates pysha3 for its new hash functions, but this feature is only in 3.6.  Thankfully, core developer tirin made pysha3 so it can be used with earlier version of python.  A compatability check uses hashlib exclusively if interpreter is >= 3.6, else, pysha3's sha3 module is imported to cover those features.  In either case, the same hash functions are used.