Added workaround for parent parser
The way parent parsers works in argparse didn't get me exactly what I wanted out of the process.  Hashchk uses a lot of argument groups so I can better control ordering and have more control.  Using parent methods was causing the order of those groups to change, and it was too hard to track where it was coming from.  It was just easier to add common arguments to their own method, and add those methods to whatever subparser is passed to that method.

The add_subparser_parsers() methods was removed.  Methods responsible for adding subparsers are called after attributes are set; add_subparser_parsers() just added an uneeded method in the end