Temporarily removed blake and shake options.  Reverted parser class to old setup
Trying to find a concise way to add blake and shake was delaying overall progress of the project and mangling the code to unmaintainable levels.  To keep things on track, their options were removed from HashChkParser and build_hash_object method in hashchk.py.  Incidentally, build_hash_method() was renamed to get_hash_method, which now returns a string instead of the method itself.  The string returned is an exact hashlib method name, but that method is now assembled in generate_digest(), which makes use of actual hashlib methods.