hashchk.py docstring additions/reformating and Digest attribute change
Added missing docstrings and reformated and/or cleaned up existing entries.  TODO sections for all sealant files reformated to fit in line with google-style docstring standards

A conditional attribute, Digest.reference_digest was added to act as a flag of sorts that the hash_method will be defined by the length of the reference_digest.  If that length doesn't match standard SHA byte lengths 56, 64, 96, or 128, the closest match is returned.  This will cause the comparison to fail, but output formatting will clearly show a length disparity allowing the user to correct their mistake.  An exception being thrown introduces too much noise for the user to easily deduce the problem areas

Changes were made to haschk_terminal.py to reflect the addition of the Digest.reference_digest parameter