Output class name change, restructure, and method changes.
Output() was renamed to Terminal() as it better suits class naming methodology.

A new method, Terminal.build_line_break() was added to automatically generate centered header messages.  This replaced numerous long print statements that held static line breaks (which were essentially a bunch of dash surrounding a title string).  While this worked, adding new prinout messages using visual line separators took up a lot of time as the length of dash lines had to be measured.  Additionally, because these print statements were so long, they were hard to manage during refactoring

The addition of build_line_break() led to a restructure of the Terminal() class as a whole.  Terminal() now operates like a normal class instead of a container for semi-related functions.  This was done so most print messages were formatted in respect to the length of the digests being printed.  For longer digests (64+ bytes), line breaks would appear uneven.  Now, 'Terminals' objects are created in respect to the size of the digests outputs the actual terminal will handle.

print_startup_message() was removed from Terminal() as build_line_break() reduced the method to one line of code.