argparse args now retrieved via RandstrParser property
Args were originally parsed by RandstrParser.get_parser_args().  In addition to returning parsed args, it called the methods responsible for adding all the arguments to the main parser; it was doing too much.  Call to add_parser_arguments is made under __init__, and parsed args are returned by args() method, which is defined as a property via @property decorator.