Proper docstrings and RandomString refactor
Expanded docstrings where detail was missing, and added docstrings if missing for all randstr files.

RandomString() class was refactored partly to allow it to function independently, and for code cleanup purposes.  Prior to this branch, RandomString() was unintentionally written to rely on RandstrParser to get its arguments.  RandomString is meant to be used by itself if desired, so some changes were needed.  Default arguments were set for some of the __init__ parameters to account for this need, for instance.

RandonString.shuffle_characters was changed from a static method to an instance method.  Instead of returning a shuffle character set, it modifies the self.char_set attribute directly.  self.generate_random_string() did this already, but this should be handled by self.shuffle_characters() directly.

RandomStrings __repr__ override was changed to __str__.  This change doesn't cause any functionality change; it was done in the name of clarity to indicate RandomString is intended to be printed.