After like 3 hours of bouncing my head off the keyboard, a semi working testing setup now exisits.
I had to use sys.path.insert() to temporarily modifiy PYTHONPATH.  Yes, it's not the best solution, but no amount of __init__.py solutions would do it.  While Pycharm will happily use a normal absolute import, the python interpreter and sublime text for that matter only respond to this method.

With that in mind, the folder structure has changed.  src is replaced with the packages name.  More changes incoming probably as trying to adhere to a professional set of standards has taught me no one can agree on what's 'best practice'.