Replaced RandomStr __str__ with __call__
The problem with what __str__ returned was that once a length parameters is provided, a string is generated every time that object is used.  While this is usually intended, if only a specific method is being called, a string is still being generated in the background.  __call__ allows for more explicit control over random string generation.  For instance, my_string = RandomString(length=10) would only return a randomly generated string in if my_string is used as my_string().

while this allows for easy creation of arbitraily named string_generator functions, the biggest benefit is a reduction of unintended side affects.

This new structure is only reflected in unittests for randstr.  Changes in other files (like RandstrParser) will be updated once this branch is finished.