setUp() method added and other test changes
setUp() method establishes an instance of RandomString along with attributes based off that instance to be used by test methods as needed.  Most testing methods were changed to make use of the setUp() method.

Method designed to make sure user character set is actually assigned was removed as it was essentially useless