New randstr_terminal parser switch, function name changes, general updates.
A new switch, --raw-output is available through RandstrParser now.  This switch surpresses all additional formatting (like visual line breaks) so output can be piped or redirected if desired.  This feature allows user to handle the randomized string in ways not covered by default options.  If this option is used, it will override all other output switches (i.e. the -p, -cp, and -f won't do anything if -ro is provided).

the main() function was renamed to randstr_output to better describe what the function does.

A new fuction, _generate_filename, replaces the default method for assigning filenames if the --file switch is used.  Previously, the default filename included the current date/time.  Explaining this in a help message made the help message it self too 'wordy', and, it wasn't a great naming convention to begin with.  _generate_filename() simply iterates through filenames formatted as 'randstr_#.txt' and returns the first iteration that isn't present in the current directory.

parser arguments used in randstr_output are provided through a paramater, parsed_args, now.  This makes the function a bit more modular, and it makes unit testing much easier.

Updated and/or removed docstrings where appropriate.