Added default character generation limit, along with switch to override that limit
Imposed a default 1000 character limit on random string generation as strings of excessive length can have negative impact on system resources.  To override this limit, a --remove-limit switch can be provided.  The RandstrOutput._generated_string property determines whether or not this switch is present and then acts accordingly.  If length value exceeding 1000 is provided, but no --remove-limit switch is present, string length defaults to 1000 characters.