hashchk.py refactor
The Digest class was refactored because some of the methods needed it, and because it was fulfilling too many purposes.  Ultimately, the Digest class is meant to determine what hash method use based on a reference digest.  The Digest class held the generate_digest() method, which wasn't exclusive to this purpose as hashchk is meant to also support simply generating a hex digest.  Furthermore, the Digest class will soon be optional as there needs to be an option that allows the user to explicitly define what hash method to use if desired.  Overall, functionality remains the same, but the code is much more organized and easier to follow over it's previous iteration.

Updated most docstrings in hashchk.py, and changed some lines of code in hashchk_terminal.py to reflect the changes in hashchk.py